requirements.txt
tqdm~=4.62.3
torch==1.10
torchvision==0.11.1
torchaudio==0.10.0
paho-mqtt==1.5.0
pika~=1.2.0
pyzmq~=22.3.0
numpy~=1.22.3